comentarios en algunos controllers + buscar productos

+++ src/main/java/riffy/model/ConversacionEntity.java
@@ -21,7 +21,7 @@ public class ConversacionEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id_conversacion;
+    private Long id_conversacion;
 
     @ManyToOne
     @JoinColumn(name = "id_producto")
@@ -45,11 +45,11 @@ public class ConversacionEntity {
     /* //////////////// getters y setters //////////////////////////////// */
     /* /////////////////////////////////////////////////////////////////// */
 
-    public Integer getId_conversacion() {
+    public Long getId_conversacion() {
         return id_conversacion;
     }
 
-    public void setId_conversacion(Integer id_conversacion) {
+    public void setId_conversacion(Long id_conversacion) {
         this.id_conversacion = id_conversacion;
     }
